Add more interfaces to block header list
